pyproject.toml
[project]
name = "buibui-moon-trader-bot"
version = "0.1.0"
description = ""
authors = [
    {name = "s10023",email = "ngkhaijian@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohappyeyeballs (>=2.6.1,<3.0.0)",
    "aiohttp (>=3.12.13,<4.0.0)",
    "aiosignal (>=1.4.0,<2.0.0)",
    "async-timeout (>=5.0.1,<6.0.0)",
    "attrs (>=25.3.0,<26.0.0)",
    "black (>=25.1.0,<26.0.0)",
    "certifi (>=2025.6.15,<2026.0.0)",
    "charset-normalizer (>=3.4.2,<4.0.0)",
    "click (>=8.2.1,<9.0.0)",
    "colorama (>=0.4.6,<0.5.0)",
    "dateparser (>=1.2.2,<2.0.0)",
    "frozenlist (>=1.7.0,<2.0.0)",
    "idna (>=3.10,<4.0)",
    "multidict (>=6.6.3,<7.0.0)",
    "mypy-extensions (>=1.1.0,<2.0.0)",
    "packaging (>=25.0,<26.0)",
    "pathspec (>=0.12.1,<0.13.0)",
    "platformdirs (>=4.3.8,<5.0.0)",
    "propcache (>=0.3.2,<0.4.0)",
    "pycryptodome (>=3.23.0,<4.0.0)",
    "python-binance (>=1.0.29,<2.0.0)",
    "python-dateutil (>=2.9.0.post0,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pytz (>=2025.2,<2026.0)",
    "regex (>=2024.11.6,<2025.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "schedule (>=1.2.2,<2.0.0)",
    "six (>=1.17.0,<2.0.0)",
    "tabulate (>=0.9.0,<0.10.0)",
    "tomli (>=2.2.1,<3.0.0)",
    "typing-extensions (>=4.14.1,<5.0.0)",
    "tzdata (>=2025.2,<2026.0)",
    "tzlocal (>=5.3.1,<6.0.0)",
    "urllib3 (>=2.5.0,<3.0.0)",
    "websockets (>=15.0.1,<16.0.0)",
    "yarl (>=1.20.1,<2.0.0)",
    "logging (>=0.4.9.6,<0.5.0.0)",
    "pytest (>=8.4.1,<9.0.0)"
]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
